Initial Commit-Recursive Dictionary
This Python Dictionary application takes input from user as a word and gives output as meaning of that word. Moreover, its ask whether the user want to get the meaning of some other word .If the user types yes then the program runs again and gives the meaning of particular word.